chore: clean baseline — drop 3 dead modules, split artifacts/__init__.py
Closes #599, partially closes #729.

## Problem

- `polylogue/artifacts/__init__.py` had grown to 1072 lines mixing type definitions
  with a large runtime artifact graph data block (#599).
- Static analysis (#729) identified 5 dead-or-test-only modules. Three are truly
  unused; two were misclassified.

## Solution

### #599 — split artifacts/__init__.py
- `polylogue/artifacts/descriptors.py` (60 lines) — `ArtifactLayer`, `ArtifactNode`, `ArtifactPath`.
- `polylogue/artifacts/runtime.py` (1005 lines) — `RUNTIME_ARTIFACT_NODES`, `RUNTIME_ARTIFACT_PATHS`, builders.
- `polylogue/artifacts/__init__.py` (21 lines) — re-export shim. All existing imports keep working.

### #729 — delete dead modules (3 of 5 candidates)
Deleted, with their tests:
- `polylogue/showcase/context.py` — confirmed dead.
- `polylogue/cli/insights_rendering.py` + `tests/unit/cli/test_insights_helper_runtime.py` — docstring already noted "replaced by generic render_insight_items()".
- `polylogue/cli/shared/qa_finalization.py` + `tests/unit/cli/test_qa_finalization.py` — never wired to a CLI verb.

Kept (with rationale):
- `polylogue/cli/shared/schema_command_support.py` — actually used by `devtools/schema_generate.py:9`. #729 missed this devtools dependency.
- `polylogue/archive/query/read_fields.py` + `tests/unit/core/test_query_read_fields.py` — load-bearing for the API-method-parameter parity check in `test_query_field_catalog_covers_api_query_method_names` (covers `content_projection`). Will be subsumed by the unified `QueryFieldDescriptor` in #621.

### Generated surface drift
Regenerated `docs/verification-catalog.md` via `python3 devtools/render_all.py`.

## Verification

- `ruff format --check polylogue/ tests/ devtools/` — clean
- `ruff check polylogue/ tests/ devtools/` — clean
- `mypy polylogue/artifacts/` — clean
- `pytest tests/unit/core/test_query_fields.py tests/unit/core/test_query_read_fields.py` — 12 passed
- Earlier full-dir pytest (1038 passed, 1 expected failure now resolved by restoring read_fields.py)
- `python -c "from polylogue.artifacts import ArtifactLayer, ArtifactNode, ArtifactPath, RUNTIME_ARTIFACT_NODES, RUNTIME_ARTIFACT_PATHS, build_runtime_artifact_nodes, build_runtime_artifact_paths"` — imports OK
