+++ polylogue/artifacts/runtime.py
@@ -1,0 +1,1005 @@
+"""Runtime artifact graph data and accessors."""
+
+from __future__ import annotations
+
+from polylogue.artifacts.descriptors import ArtifactLayer, ArtifactNode, ArtifactPath
+
+RUNTIME_ARTIFACT_NODES: tuple[ArtifactNode, ...] = (
+    ArtifactNode(
+        name="configured_sources",
+        layer=ArtifactLayer.SOURCE,
+        description="Configured filesystem and Drive source descriptors selected for an archive run.",
+        code_refs=(
+            "polylogue.config.Source",
+            "polylogue.pipeline.run_planning.plan_sources",
+            "polylogue.cli.shared.helpers.resolve_sources",
+        ),
+        readiness_surfaces=("run", "sources"),
+    ),
+    ArtifactNode(
+        name="source_payload_stream",
+        layer=ArtifactLayer.SOURCE,
+        description="Traversed raw source payloads yielded as RawConversationData before durable persistence.",
+        depends_on=("configured_sources",),
+        code_refs=(
+            "polylogue.sources.source_acquisition.iter_source_raw_data",
+            "polylogue.pipeline.services.acquisition.AcquisitionService.visit_sources",
+        ),
+        readiness_surfaces=("run", "sources"),
+    ),
+    ArtifactNode(
+        name="raw_validation_state",
+        layer=ArtifactLayer.DURABLE,
+        description="Persisted raw-conversation validation and parse state in raw_conversations.",
+        depends_on=("source_payload_stream",),
+        code_refs=(
+            "polylogue.storage.raw.artifacts.RawIngestArtifactState",
+            "polylogue.storage.backends.queries.raw_state",
+        ),
+    ),
+    ArtifactNode(
+        name="artifact_observation_rows",
+        layer=ArtifactLayer.DURABLE,
+        description="Persisted artifact observations recorded alongside acquired raw conversations.",
+        depends_on=("raw_validation_state",),
+        code_refs=(
+            "polylogue.pipeline.services.acquisition_persistence.persist_raw_record",
+            "polylogue.storage.repository.archive.repository_writes.RepositoryWritesMixin.save_artifact_observation",
+            "polylogue.storage.artifacts.inspection.inspect_raw_artifact",
+        ),
+        readiness_surfaces=("run", "sources"),
+    ),
+    ArtifactNode(
+        name="validation_backlog",
+        layer=ArtifactLayer.PROJECTION,
+        description="Raw records that still require validation before ordinary parse planning can trust them.",
+        depends_on=("raw_validation_state",),
+        code_refs=(
+            "polylogue.storage.raw.artifacts.validation_backlog_query_spec",
+            "polylogue.pipeline.services.planning_backlog.collect_validation_backlog",
+        ),
+    ),
+    ArtifactNode(
+        name="parse_backlog",
+        layer=ArtifactLayer.PROJECTION,
+        description="Raw records eligible for parse planning, including force-reparse simulation.",
+        depends_on=("raw_validation_state",),
+        code_refs=(
+            "polylogue.storage.raw.artifacts.parse_backlog_query_spec",
+            "polylogue.pipeline.services.planning_backlog.collect_parse_backlog",
+            "polylogue.pipeline.services.planning_runtime.build_ingest_plan",
+        ),
+    ),
+    ArtifactNode(
+        name="parse_quarantine",
+        layer=ArtifactLayer.PROJECTION,
+        description="Validation-failed unparsed raws that stay out of ordinary parse backlog but return under force-reparse.",
+        depends_on=("raw_validation_state",),
+        code_refs=("polylogue.storage.raw.artifacts.RawIngestArtifactState",),
+    ),
+    ArtifactNode(
+        name="archive_conversation_rows",
+        layer=ArtifactLayer.DURABLE,
+        description="Durable conversation, message, attachment, and content-block rows written from parsed raw conversations.",
+        depends_on=("raw_validation_state",),
+        code_refs=(
+            "polylogue.pipeline.services.parsing_workflow.parse_from_raw",
+            "polylogue.pipeline.prepare.prepare_records",
+            "polylogue.pipeline.prepare.persist_prepared_bundle",
+            "polylogue.storage.repository.archive.writes.conversations.save_via_backend",
+            "polylogue.storage.repository.raw.repository_raw.RepositoryRawMixin.mark_raw_parsed",
+        ),
+    ),
+    ArtifactNode(
+        name="message_source_rows",
+        layer=ArtifactLayer.SOURCE,
+        description="Persisted message rows that feed lexical FTS indexing and archive search.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.fts.fts_lifecycle.FTS_INDEXABLE_MESSAGE_COUNT_SQL",
+            "polylogue.storage.fts.fts_lifecycle.repair_fts_index_sync",
+        ),
+    ),
+    ArtifactNode(
+        name="message_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Lexical FTS projection over persisted message rows.",
+        depends_on=("message_source_rows",),
+        code_refs=(
+            "polylogue.storage.fts.fts_lifecycle.message_fts_readiness_sync",
+            "polylogue.storage.fts.fts_lifecycle.rebuild_fts_index_sync",
+            "polylogue.storage.fts.fts_lifecycle.repair_fts_index_sync",
+        ),
+        repair_targets=("dangling_fts",),
+        readiness_surfaces=("doctor", "archive_debt", "query"),
+    ),
+    ArtifactNode(
+        name="embedding_metadata_rows",
+        layer=ArtifactLayer.DURABLE,
+        description="Durable embedding metadata rows storing model, dimension, timestamp, and provenance for embedded messages.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.search_providers.sqlite_vec_runtime.SqliteVecRuntimeMixin._ensure_tables",
+            "polylogue.storage.search_providers.sqlite_vec_queries.SqliteVecQueryMixin.upsert",
+        ),
+        readiness_surfaces=("embed", "doctor", "retrieval"),
+    ),
+    ArtifactNode(
+        name="embedding_status_rows",
+        layer=ArtifactLayer.DURABLE,
+        description="Durable per-conversation embedding status rows recording embedded message counts and reindex needs.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.search_providers.sqlite_vec_runtime.SqliteVecRuntimeMixin._ensure_tables",
+            "polylogue.storage.search_providers.sqlite_vec_queries.SqliteVecQueryMixin.upsert",
+        ),
+        readiness_surfaces=("embed", "doctor", "retrieval"),
+    ),
+    ArtifactNode(
+        name="message_embedding_vectors",
+        layer=ArtifactLayer.INDEX,
+        description="Semantic vector index over embedded archive messages.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.search_providers.sqlite_vec_runtime.SqliteVecRuntimeMixin._ensure_tables",
+            "polylogue.storage.search_providers.sqlite_vec_queries.SqliteVecQueryMixin.upsert",
+            "polylogue.storage.search_providers.sqlite_vec_queries.SqliteVecQueryMixin.query",
+        ),
+        readiness_surfaces=("embed", "retrieval", "query"),
+    ),
+    ArtifactNode(
+        name="tool_use_source_blocks",
+        layer=ArtifactLayer.SOURCE,
+        description="Tool-use content blocks anchored to valid conversations.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=("polylogue.storage.action_events.status",),
+    ),
+    ArtifactNode(
+        name="action_event_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Materialized action-event read model derived from tool-use source blocks.",
+        depends_on=("tool_use_source_blocks",),
+        code_refs=(
+            "polylogue.storage.action_events.artifacts.ActionEventArtifactState",
+            "polylogue.storage.action_events.status",
+        ),
+        repair_targets=("action_event_read_model",),
+        readiness_surfaces=("doctor", "archive_debt"),
+    ),
+    ArtifactNode(
+        name="action_event_fts",
+        layer=ArtifactLayer.INDEX,
+        description="FTS projection over the action-event read model.",
+        depends_on=("action_event_rows",),
+        code_refs=(
+            "polylogue.storage.action_events.artifacts.ActionEventArtifactState",
+            "polylogue.storage.derived.insights.build_action_statuses",
+        ),
+        repair_targets=("action_event_read_model",),
+        readiness_surfaces=("doctor", "archive_debt", "retrieval_evidence"),
+    ),
+    ArtifactNode(
+        name="action_event_readiness",
+        layer=ArtifactLayer.PROJECTION,
+        description="Projected readiness, debt, and repair semantics over action-event rows and FTS.",
+        depends_on=("action_event_rows", "action_event_fts"),
+        code_refs=(
+            "polylogue.storage.derived.derived_status",
+            "polylogue.storage.repair",
+            "polylogue.storage.embeddings.support",
+        ),
+        repair_targets=("action_event_read_model",),
+        readiness_surfaces=("doctor", "archive_debt", "retrieval_evidence"),
+    ),
+    ArtifactNode(
+        name="session_insight_source_conversations",
+        layer=ArtifactLayer.SOURCE,
+        description="Hydratable conversation/message/attachment/block rows that feed durable session-insight rebuilds.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.rebuild.rebuild_session_insights_sync",
+            "polylogue.storage.insights.session.refresh.refresh_session_insights_for_conversation_async",
+        ),
+    ),
+    ArtifactNode(
+        name="session_profile_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable session profile rows materialized from archive conversations.",
+        depends_on=("session_insight_source_conversations",),
+        code_refs=(
+            "polylogue.storage.insights.session.rebuild.build_session_insight_records",
+            "polylogue.storage.insights.session.profiles.build_session_profile_record",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_profile_merged_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Merged session-profile FTS projection over durable session profile rows.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_profiles",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_profile_evidence_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Evidence-tier session-profile FTS projection over durable session profile rows.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_profiles",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_profile_inference_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Inference-tier session-profile FTS projection over durable session profile rows.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_profiles",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_profile_enrichment_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Enrichment-tier session-profile FTS projection over durable session profile rows.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_profiles",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_work_event_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable session work-event rows materialized from session profile analysis.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.rebuild.build_session_insight_records",
+            "polylogue.storage.insights.session.timeline_rows.build_session_work_event_records",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_work_event_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Session work-event FTS projection over durable work-event rows.",
+        depends_on=("session_work_event_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_timelines",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_phase_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable session phase rows materialized from session profile analysis.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.rebuild.build_session_insight_records",
+            "polylogue.storage.insights.session.timeline_rows.build_session_phase_records",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="work_thread_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable work-thread rows materialized from session profile families.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.threads.build_thread_records_for_roots_sync",
+            "polylogue.storage.insights.session.threads.build_thread_records_for_roots_async",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="work_thread_fts",
+        layer=ArtifactLayer.INDEX,
+        description="Work-thread FTS projection over durable work-thread rows.",
+        depends_on=("work_thread_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.backends.schema_ddl_insight_aggregates",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_tag_rollup_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable provider/day/tag rollup rows aggregated from session profiles.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.aggregates.refresh_sync_provider_day_aggregates",
+            "polylogue.storage.insights.session.status",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="day_session_summary_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable provider/day summary rows aggregated from session profiles.",
+        depends_on=("session_profile_rows",),
+        code_refs=(
+            "polylogue.storage.insights.session.aggregates.refresh_sync_provider_day_aggregates",
+            "polylogue.storage.insights.session.status",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_insight_rows",
+        layer=ArtifactLayer.DERIVED,
+        description="Durable session-profile, timeline, thread, and aggregate rows derived from archive conversations.",
+        depends_on=(
+            "session_profile_rows",
+            "session_work_event_rows",
+            "session_phase_rows",
+            "work_thread_rows",
+            "session_tag_rollup_rows",
+            "day_session_summary_rows",
+        ),
+        code_refs=(
+            "polylogue.storage.insights.session.rebuild.rebuild_session_insights_sync",
+            "polylogue.storage.insights.session.status",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_insight_fts",
+        layer=ArtifactLayer.INDEX,
+        description="The session-insight FTS family over profiles, work events, and threads.",
+        depends_on=(
+            "session_profile_merged_fts",
+            "session_profile_evidence_fts",
+            "session_profile_inference_fts",
+            "session_profile_enrichment_fts",
+            "session_work_event_fts",
+            "work_thread_fts",
+        ),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.derived.insights",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="session_insight_readiness",
+        layer=ArtifactLayer.PROJECTION,
+        description="Projected readiness, debt, and stale-surface semantics for durable session insights.",
+        depends_on=("session_insight_rows", "session_insight_fts"),
+        code_refs=(
+            "polylogue.storage.insights.session.status",
+            "polylogue.storage.repair",
+            "polylogue.cli.commands.insights",
+        ),
+        repair_targets=("session_insights",),
+        readiness_surfaces=("doctor", "archive_debt", "insights"),
+    ),
+    ArtifactNode(
+        name="retrieval_band_readiness",
+        layer=ArtifactLayer.PROJECTION,
+        description="Projected transcript/evidence/inference/enrichment retrieval readiness over embeddings and durable read models.",
+        depends_on=(
+            "embedding_metadata_rows",
+            "embedding_status_rows",
+            "message_embedding_vectors",
+            "action_event_readiness",
+            "session_insight_readiness",
+        ),
+        code_refs=(
+            "polylogue.storage.embeddings.embedding_stats.read_embedding_stats_sync",
+            "polylogue.storage.embeddings.support.build_retrieval_bands_from_status",
+            "polylogue.storage.derived.derived_status.build_retrieval_statuses",
+        ),
+        readiness_surfaces=("embed", "doctor", "retrieval"),
+    ),
+    ArtifactNode(
+        name="embedding_status_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Operator-facing embedding status payload and retrieval-band readiness view.",
+        depends_on=(
+            "embedding_metadata_rows",
+            "embedding_status_rows",
+            "message_embedding_vectors",
+            "retrieval_band_readiness",
+        ),
+        code_refs=(
+            "polylogue.storage.embeddings.status_payload.embedding_status_payload",
+            "polylogue.cli.shared.embed_stats.show_embedding_stats",
+        ),
+        readiness_surfaces=("embed", "doctor", "retrieval", "cli"),
+    ),
+    ArtifactNode(
+        name="session_profile_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session profiles.",
+        depends_on=("session_profile_rows", "session_profile_merged_fts"),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_session_profile_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="session_enrichment_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session enrichments.",
+        depends_on=("session_profile_rows", "session_profile_enrichment_fts"),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_session_enrichment_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="session_work_event_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session work events.",
+        depends_on=("session_work_event_rows", "session_work_event_fts"),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_session_work_event_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="session_phase_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session phases.",
+        depends_on=("session_phase_rows",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_session_phase_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="work_thread_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable work threads.",
+        depends_on=("work_thread_rows", "work_thread_fts"),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_work_thread_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="session_tag_rollup_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session tag rollups.",
+        depends_on=("session_tag_rollup_rows",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_session_tag_rollup_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="day_session_summary_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable day session summaries.",
+        depends_on=("day_session_summary_rows",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_day_session_summary_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="week_session_summary_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for week session summaries derived from durable day summaries.",
+        depends_on=("day_session_summary_rows",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_week_session_summary_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="session_insight_status_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for durable session-insight status views.",
+        depends_on=("session_insight_readiness",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveStatsMixin.get_session_insight_status",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp"),
+    ),
+    ArtifactNode(
+        name="provider_analytics_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for provider analytics derived from durable session insights.",
+        depends_on=("session_insight_rows",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightMixin.list_provider_analytics_insights",
+            "polylogue.cli.commands.insights",
+            "polylogue.cli.shared.helper_summary",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp", "helpers"),
+    ),
+    ArtifactNode(
+        name="archive_debt_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Query/read results for archive debt views derived from projected readiness and maintenance state.",
+        depends_on=("action_event_readiness", "session_insight_readiness", "archive_readiness"),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveInsightDebtMixin.list_archive_debt_insights",
+            "polylogue.cli.commands.insights",
+        ),
+        readiness_surfaces=("insights", "facade", "mcp", "maintenance"),
+    ),
+    ArtifactNode(
+        name="conversation_query_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Conversation-level query and search results resolved from lexical retrieval over the archive.",
+        depends_on=("message_fts",),
+        code_refs=(
+            "polylogue.operations.archive.ArchiveSearchMixin.query_conversations",
+            "polylogue.operations.archive.ArchiveSearchMixin.search",
+            "polylogue.archive.query.plan_execution",
+        ),
+        readiness_surfaces=("query", "mcp", "facade"),
+    ),
+    ArtifactNode(
+        name="conversation_render_projection",
+        layer=ArtifactLayer.PROJECTION,
+        description="Repository-owned render projection preserving canonical conversation, message, and attachment layout.",
+        depends_on=("archive_conversation_rows",),
+        code_refs=(
+            "polylogue.storage.repository.archive.conversations.RepositoryArchiveConversationMixin.get_render_projection",
+            "polylogue.rendering.core_formatter.ConversationFormatter.load_projection",
+        ),
+        readiness_surfaces=("render", "site"),
+    ),
+    ArtifactNode(
+        name="rendered_conversation_artifacts",
+        layer=ArtifactLayer.PROJECTION,
+        description="Filesystem render outputs produced by run render for conversation pages.",
+        depends_on=("conversation_render_projection",),
+        code_refs=(
+            "polylogue.pipeline.services.rendering.RenderService.render_conversations",
+            "polylogue.rendering.renderers.markdown.MarkdownRenderer.render",
+            "polylogue.rendering.renderers.html.HTMLRenderer.render",
+        ),
+        readiness_surfaces=("render",),
+    ),
+    ArtifactNode(
+        name="site_conversation_pages",
+        layer=ArtifactLayer.PROJECTION,
+        description="Conversation page artifacts emitted by the static-site builder.",
+        depends_on=("conversation_render_projection",),
+        code_refs=(
+            "polylogue.site.conversation_pages.generate_conversation_page",
+            "polylogue.site.site_builder_archive.generate_conversation_page_for_builder",
+        ),
+        readiness_surfaces=("site",),
+    ),
+    ArtifactNode(
+        name="site_publication_manifest",
+        layer=ArtifactLayer.PROJECTION,
+        description="Typed site publication manifest written to disk after a successful site build.",
+        depends_on=("site_conversation_pages",),
+        code_refs=(
+            "polylogue.site.publication_flow.build_site_publication_manifest",
+            "polylogue.site.publication_flow.write_site_publication_manifest",
+        ),
+        readiness_surfaces=("site", "maintenance"),
+    ),
+    ArtifactNode(
+        name="publication_records",
+        layer=ArtifactLayer.DURABLE,
+        description="Persisted publication records storing the latest publication manifests in the archive database.",
+        depends_on=("site_publication_manifest",),
+        code_refs=(
+            "polylogue.storage.repository.archive.repository_writes.RepositoryWritesMixin.record_publication",
+            "polylogue.storage.repository.archive.repository_writes.RepositoryWritesMixin.get_latest_publication",
+            "polylogue.storage.backends.queries.publications",
+        ),
+        readiness_surfaces=("site", "maintenance"),
+    ),
+    ArtifactNode(
+        name="schema_packages",
+        layer=ArtifactLayer.DURABLE,
+        description="Versioned provider schema packages stored in the schema registry.",
+        code_refs=(
+            "polylogue.schemas.registry.SchemaRegistry",
+            "polylogue.schemas.operator.inference.list_schemas",
+            "polylogue.schemas.operator.resolution.explain_schema",
+        ),
+        readiness_surfaces=("schema", "qa"),
+    ),
+    ArtifactNode(
+        name="schema_cluster_manifests",
+        layer=ArtifactLayer.DURABLE,
+        description="Persisted schema-cluster manifests describing observed structural families per provider.",
+        code_refs=(
+            "polylogue.schemas.registry.SchemaRegistry",
+            "polylogue.schemas.operator.inference.list_schemas",
+        ),
+        readiness_surfaces=("schema", "qa"),
+    ),
+    ArtifactNode(
+        name="inferred_corpus_specs",
+        layer=ArtifactLayer.PROJECTION,
+        description="Synthetic corpus specs distilled from schema packages and cluster manifests.",
+        depends_on=("schema_packages", "schema_cluster_manifests"),
+        code_refs=(
+            "polylogue.scenarios.corpus.build_inferred_corpus_specs",
+            "polylogue.schemas.operator.inference.list_inferred_corpus_specs",
+        ),
+        readiness_surfaces=("schema", "qa", "synthetic"),
+    ),
+    ArtifactNode(
+        name="inferred_corpus_scenarios",
+        layer=ArtifactLayer.PROJECTION,
+        description="Compiled inferred corpus scenarios grouped from inferred corpus specs.",
+        depends_on=("inferred_corpus_specs",),
+        code_refs=(
+            "polylogue.scenarios.corpus.build_corpus_scenarios",
+            "polylogue.schemas.operator.inference.list_inferred_corpus_scenarios",
+        ),
+        readiness_surfaces=("schema", "qa", "synthetic"),
+    ),
+    ArtifactNode(
+        name="schema_list_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Schema list/read results combining package catalogs, manifests, and inferred corpus projections.",
+        depends_on=(
+            "schema_packages",
+            "schema_cluster_manifests",
+            "inferred_corpus_specs",
+            "inferred_corpus_scenarios",
+        ),
+        code_refs=(
+            "polylogue.schemas.operator.inference.list_schemas",
+            "polylogue.cli.commands.schema.schema_list",
+        ),
+        readiness_surfaces=("schema", "cli"),
+    ),
+    ArtifactNode(
+        name="schema_explanation_results",
+        layer=ArtifactLayer.PROJECTION,
+        description="Schema explain/read results for provider package elements and annotations.",
+        depends_on=("schema_packages",),
+        code_refs=(
+            "polylogue.schemas.operator.resolution.explain_schema",
+            "polylogue.cli.commands.schema.schema_explain",
+        ),
+        readiness_surfaces=("schema", "cli"),
+    ),
+    ArtifactNode(
+        name="archive_readiness",
+        layer=ArtifactLayer.PROJECTION,
+        description="Projected archive-wide readiness and maintenance view over message FTS and durable derived-model readiness.",
+        depends_on=("message_fts", "action_event_readiness", "session_insight_readiness", "retrieval_band_readiness"),
+        code_refs=(
+            "polylogue.readiness.run_archive_readiness",
+            "polylogue.storage.derived.derived_status.collect_derived_model_statuses_sync",
+            "polylogue.storage.repair.collect_archive_debt_statuses_sync",
+        ),
+        repair_targets=("dangling_fts", "action_event_read_model", "session_insights"),
+        readiness_surfaces=("doctor", "archive_debt", "maintenance"),
+    ),
+)
+
+RUNTIME_ARTIFACT_PATHS: tuple[ArtifactPath, ...] = (
+    ArtifactPath(
+        name="source-acquisition-loop",
+        description="Configured sources through traversed raw payloads into durable raw conversation state and artifact observations.",
+        nodes=(
+            "configured_sources",
+            "source_payload_stream",
+            "raw_validation_state",
+            "artifact_observation_rows",
+        ),
+    ),
+    ArtifactPath(
+        name="raw-reparse-loop",
+        description="Raw validation state to validation/parse backlog and quarantine projections.",
+        nodes=(
+            "raw_validation_state",
+            "validation_backlog",
+            "parse_backlog",
+            "parse_quarantine",
+        ),
+    ),
+    ArtifactPath(
+        name="raw-archive-ingest-loop",
+        description="Raw validation and parse planning through durable archive runtime rows and their downstream source surfaces.",
+        nodes=(
+            "configured_sources",
+            "source_payload_stream",
+            "raw_validation_state",
+            "artifact_observation_rows",
+            "validation_backlog",
+            "parse_backlog",
+            "parse_quarantine",
+            "archive_conversation_rows",
+            "message_source_rows",
+            "tool_use_source_blocks",
+            "session_insight_source_conversations",
+        ),
+    ),
+    ArtifactPath(
+        name="action-event-repair-loop",
+        description="Tool-use source blocks through action-event rows, FTS, and projected repair semantics.",
+        nodes=(
+            "tool_use_source_blocks",
+            "action_event_rows",
+            "action_event_fts",
+            "action_event_readiness",
+        ),
+    ),
+    ArtifactPath(
+        name="embedding-materialization-loop",
+        description="Durable archive conversations through embedding metadata, status rows, and vector index materialization.",
+        nodes=(
+            "archive_conversation_rows",
+            "embedding_metadata_rows",
+            "embedding_status_rows",
+            "message_embedding_vectors",
+        ),
+    ),
+    ArtifactPath(
+        name="session-insight-repair-loop",
+        description="Archive conversations through durable session-insight rows, FTS, and projected repair semantics.",
+        nodes=(
+            "session_insight_source_conversations",
+            "session_profile_rows",
+            "session_profile_merged_fts",
+            "session_profile_evidence_fts",
+            "session_profile_inference_fts",
+            "session_profile_enrichment_fts",
+            "session_work_event_rows",
+            "session_work_event_fts",
+            "session_phase_rows",
+            "work_thread_rows",
+            "work_thread_fts",
+            "session_tag_rollup_rows",
+            "day_session_summary_rows",
+            "session_insight_rows",
+            "session_insight_fts",
+            "session_insight_readiness",
+        ),
+    ),
+    ArtifactPath(
+        name="raw-session-insight-repair-loop",
+        description=(
+            "Raw validation and archive core rows through durable session-insight rows, FTS, and projected repair semantics."
+        ),
+        nodes=(
+            "configured_sources",
+            "source_payload_stream",
+            "raw_validation_state",
+            "archive_conversation_rows",
+            "session_insight_source_conversations",
+            "session_profile_rows",
+            "session_work_event_rows",
+            "session_phase_rows",
+            "work_thread_rows",
+            "session_tag_rollup_rows",
+            "day_session_summary_rows",
+            "session_insight_rows",
+            "session_insight_fts",
+            "session_insight_readiness",
+        ),
+    ),
+    ArtifactPath(
+        name="retrieval-band-readiness-loop",
+        description="Embedding state plus action/session-insight readiness through retrieval-band and archive readiness projections.",
+        nodes=(
+            "embedding_metadata_rows",
+            "embedding_status_rows",
+            "message_embedding_vectors",
+            "action_event_readiness",
+            "session_insight_readiness",
+            "retrieval_band_readiness",
+            "archive_readiness",
+        ),
+    ),
+    ArtifactPath(
+        name="session-profile-query-loop",
+        description="Durable session profile rows through profile FTS and profile query results.",
+        nodes=(
+            "session_profile_rows",
+            "session_profile_merged_fts",
+            "session_profile_results",
+        ),
+    ),
+    ArtifactPath(
+        name="session-enrichment-query-loop",
+        description="Durable session profile rows through enrichment FTS and enrichment query results.",
+        nodes=(
+            "session_profile_rows",
+            "session_profile_enrichment_fts",
+            "session_enrichment_results",
+        ),
+    ),
+    ArtifactPath(
+        name="session-work-event-query-loop",
+        description="Durable session work-event rows through work-event FTS and query results.",
+        nodes=(
+            "session_work_event_rows",
+            "session_work_event_fts",
+            "session_work_event_results",
+        ),
+    ),
+    ArtifactPath(
+        name="session-phase-query-loop",
+        description="Durable session phase rows through phase query results.",
+        nodes=(
+            "session_phase_rows",
+            "session_phase_results",
+        ),
+    ),
+    ArtifactPath(
+        name="work-thread-query-loop",
+        description="Durable work-thread rows through work-thread FTS and query results.",
+        nodes=(
+            "work_thread_rows",
+            "work_thread_fts",
+            "work_thread_results",
+        ),
+    ),
+    ArtifactPath(
+        name="session-tag-rollup-query-loop",
+        description="Durable tag-rollup rows through tag-rollup query results.",
+        nodes=(
+            "session_tag_rollup_rows",
+            "session_tag_rollup_results",
+        ),
+    ),
+    ArtifactPath(
+        name="day-summary-query-loop",
+        description="Durable day-summary rows through day-summary query results.",
+        nodes=(
+            "day_session_summary_rows",
+            "day_session_summary_results",
+        ),
+    ),
+    ArtifactPath(
+        name="week-summary-query-loop",
+        description="Durable day-summary rows through week-summary query results.",
+        nodes=(
+            "day_session_summary_rows",
+            "week_session_summary_results",
+        ),
+    ),
+    ArtifactPath(
+        name="session-insight-status-query-loop",
+        description="Projected session-insight readiness through status query results.",
+        nodes=(
+            "session_insight_readiness",
+            "session_insight_status_results",
+        ),
+    ),
+    ArtifactPath(
+        name="provider-analytics-query-loop",
+        description="Durable session-insight aggregates through provider-analytics query results.",
+        nodes=(
+            "session_insight_rows",
+            "provider_analytics_results",
+        ),
+    ),
+    ArtifactPath(
+        name="embedding-status-query-loop",
+        description="Embedding state and retrieval-band readiness through operator-facing embedding status results.",
+        nodes=(
+            "embedding_metadata_rows",
+            "embedding_status_rows",
+            "message_embedding_vectors",
+            "retrieval_band_readiness",
+            "embedding_status_results",
+        ),
+    ),
+    ArtifactPath(
+        name="archive-debt-query-loop",
+        description="Projected derived-model readiness through archive debt query results.",
+        nodes=(
+            "action_event_readiness",
+            "session_insight_readiness",
+            "archive_readiness",
+            "archive_debt_results",
+        ),
+    ),
+    ArtifactPath(
+        name="message-fts-readiness-loop",
+        description="Persisted messages through lexical FTS and the archive-wide readiness projection.",
+        nodes=(
+            "message_source_rows",
+            "message_fts",
+            "action_event_readiness",
+            "session_insight_readiness",
+            "archive_readiness",
+        ),
+    ),
+    ArtifactPath(
+        name="conversation-query-loop",
+        description="Lexical message FTS through conversation-level query and search result projections.",
+        nodes=(
+            "message_fts",
+            "conversation_query_results",
+        ),
+    ),
+    ArtifactPath(
+        name="conversation-render-loop",
+        description="Durable archive conversations through repository render projections and rendered filesystem artifacts.",
+        nodes=(
+            "archive_conversation_rows",
+            "conversation_render_projection",
+            "rendered_conversation_artifacts",
+        ),
+    ),
+    ArtifactPath(
+        name="site-publication-loop",
+        description="Durable archive conversations through site conversation pages, publication manifest output, and persisted publication records.",
+        nodes=(
+            "archive_conversation_rows",
+            "conversation_render_projection",
+            "site_conversation_pages",
+            "site_publication_manifest",
+            "publication_records",
+        ),
+    ),
+    ArtifactPath(
+        name="inferred-corpus-compilation-loop",
+        description="Schema packages and cluster manifests through inferred corpus specs and compiled inferred corpus scenarios.",
+        nodes=(
+            "schema_packages",
+            "schema_cluster_manifests",
+            "inferred_corpus_specs",
+            "inferred_corpus_scenarios",
+        ),
+    ),
+    ArtifactPath(
+        name="schema-list-query-loop",
+        description="Schema registry packages and manifests through inferred corpus projections and schema list results.",
+        nodes=(
+            "schema_packages",
+            "schema_cluster_manifests",
+            "inferred_corpus_specs",
+            "inferred_corpus_scenarios",
+            "schema_list_results",
+        ),
+    ),
+    ArtifactPath(
+        name="schema-explain-query-loop",
+        description="Schema registry packages through schema explanation results.",
+        nodes=(
+            "schema_packages",
+            "schema_explanation_results",
+        ),
+    ),
+)
+
+
+def build_runtime_artifact_nodes() -> tuple[ArtifactNode, ...]:
+    return RUNTIME_ARTIFACT_NODES
+
+
+def build_runtime_artifact_paths() -> tuple[ArtifactPath, ...]:
+    return RUNTIME_ARTIFACT_PATHS
